requirements.txt
orxhestra>=0.0.50
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
pytest>=7.0.0